Removed IBoard interface alltogether
The interface was no longer needed as the external board is now
managed by the internal board. This means that it doesn't have to
have a consistent interface with the internal board, as a player
(the user) is no longer calling both boards. The interfaces of
the external and internal boards can now be different to better
support the promotion logic.